Trim when splitting strings from Google Doc
