+++ static/src/javascripts/projects/common/modules/experiments/tests/acquisitions-banner-australia-post-one-million.js
@@ -1,7 +1,7 @@
 // @flow strict
 import { makeBannerABTestVariants } from 'common/modules/commercial/contributions-utilities';
 import { getSync as geolocationGetSync } from 'lib/geolocation';
-import { makeBannerABTestVariant } from 'common/modules/commercial/membership-engagement-banner-parameters';
+import { makeBannerABTestVariant } from 'common/modules/commercial/contributions-utilities';
 
 const componentType: OphanComponentType = 'ACQUISITIONS_ENGAGEMENT_BANNER';
 const abTestName = 'AcquisitionsBannerAustraliaPostOneMillion';
